almost done

+++ src/tasks/tasks.service.ts
@@ -16,8 +16,10 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ) { }
 
-    getTasks(filterDto: GetTasksFilterDto) {
-        return this.taskRepository.getTasks(filterDto);
+    getTasks(
+        filterDto: GetTasksFilterDto,
+        user: User) {
+        return this.taskRepository.getTasks(filterDto, user);
     }
     // private tasks: Task[] = [];
 
@@ -42,8 +44,10 @@ export class TasksService {
     //     return tasks;
     // }
 
-    async getTaskById(id: number): Promise<Task> {
-        const found = await this.taskRepository.findOne(id);
+    async getTaskById(
+        id: number,
+        user: User): Promise<Task> {
+        const found = await this.taskRepository.findOne({where: {id, userId: user.id}});
 
         if (!found) {
             throw new NotFoundException(`Task with ID "${id}" Not found`);
@@ -78,8 +82,8 @@ export class TasksService {
     //     return task;
     // }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id});
         if (result.affected === 0) {
             throw new NotFoundException(`Task with ID "${id}" Not found`);
         }
@@ -89,8 +93,8 @@ export class TasksService {
     //     this.tasks = this.tasks.filter(tasks => tasks.id !== found.id);
     // }
 
-    async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-       const task = await this.getTaskById(id);
+    async updateTaskStatus(id: number, status: TaskStatus, user: User): Promise<Task> {
+       const task = await this.getTaskById(id, user);
        task.status = status;
        await task.save();
        return task;
